requirements.txt
feedparser
requests
beautifulsoup4
readability-lxml
python-dateutil
pyyaml
sentence-transformers

